# gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: WaaV Gateway API
  description: Real-time voice processing server with Speech-to-Text (STT) and Text-to-Speech (TTS) services
  contact:
    name: WaaV
    url: https://api.waav.ai
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: https://api.waav.ai
  description: Production API
- url: http://localhost:3001
  description: Local development
paths:
...
  /recording/{stream_id}:
    get:
      tags:
      - recordings
      summary: Download recording by stream ID from configured object storage
      description: |-
        When authentication is enabled, recordings are scoped to the authenticated
        client's ID. Users can only download recordings that belong to their tenant.
        The recording path includes auth_id prefix: `{prefix}/{auth_id}/{stream_id}/audio.ogg`
      operationId: download_recording
      parameters:
      - name: stream_id
        in: path
        description: Recording stream identifier
        required: true
        schema:
          type: string
        example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Recording retrieved successfully
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Suggested filename for download
            Content-Length:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Size of the recording in bytes
          content:
            audio/ogg: {}
        '400':
          description: Invalid stream_id format
        '404':
          description: Recording not found
        '503':
          description: Recording storage not configured or unavailable
      security:
      - bearer_auth: []
...
components:
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication service. Required when AUTH_REQUIRED is enabled.
...